refactor: package 구조 클린아키텍쳐로 변경
