+++ app/src/main/java/com/survivalcoding/stopwatch/presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.survivalcoding.stopwatch.activity
+package com.survivalcoding.stopwatch.presentation
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
@@ -7,8 +7,8 @@ import androidx.fragment.app.commit
 import androidx.fragment.app.replace
 import com.survivalcoding.stopwatch.R
 import com.survivalcoding.stopwatch.databinding.ActivityMainBinding
-import com.survivalcoding.stopwatch.fragment.BlankFragment
-import com.survivalcoding.stopwatch.fragment.StopWatchFragment
+import com.survivalcoding.stopwatch.presentation.blank.BlankFragment
+import com.survivalcoding.stopwatch.presentation.stopwatch.StopWatchFragment
 
 class MainActivity : AppCompatActivity() {
