+++ app/src/main/res/layout-land/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/light_black"
 
-    tools:context=".activity.MainActivity">
+    tools:context=".presentation.MainActivity">
 
 
     <FrameLayout
