+++ app/src/main/res/layout-land/fragment_stop_watch.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragment.StopWatchFragment">
+    tools:context=".presentation.stopwatch.StopWatchFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
